Polish, address some issues
