+++ app/contribute/content.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { MotionContainer } from "@/components/container"
+import { Container } from "@/components/container"
 import ScrollSection from "@/components/scroll-section"
 import Cta from "./cta"
 import Function from "./function"
@@ -9,11 +9,7 @@ import Intro from "./intro"
 
 export default function Contribute() {
   return (
-    <MotionContainer
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      exit={{ opacity: 0 }}
-    >
+    <Container>
       <ScrollSection height="200vh">
         {percentage => <Header percentage={percentage} />}
       </ScrollSection>
@@ -22,6 +18,6 @@ export default function Contribute() {
         {percentage => <Function percentage={percentage} />}
       </ScrollSection>
       <Cta />
-    </MotionContainer>
+    </Container>
   )
 }
