+++ public/robots.txt
@@ -2,4 +2,4 @@ User-agent: Twitterbot
 Disallow:
 
 User-agent: *
-Disallow: /
+Disallow:
